6-size.c

+++ 0x00-hello_world/6-size.c
@@ -12,11 +12,11 @@ int main(void)
 	long long int d;
 	float f;
 
-printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-printf("size of int: %lu byte(s)\n", (unsigned long)sizeof(b));
-printf("size of long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("size of long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-printf("size of float: %lu byte(s)\n", (unsigned long)sizeof(f));
+printf("size of a char: %zu byte(s)\n", sizeof(a));
+printf("size of int: %zu byte(s)\n", sizeof(b));
+printf("size of long int: %zu byte(s)\n", sizeof(c));
+printf("size of long long int: %zu byte(s)\n", sizeof(d));
+printf("size of float: %zu byte(s)\n", sizeof(f));
 
 return (0);
 }
